Upload cache fixes
This mod fixes the following upload cache problems:

(1)
If one uploads many files (dropping files test0.bin-test9.bin in one move) races may occur and file id in DOM won't match file ids in web upload dir on server, i.e.:

```
$(".AttachmentListContainer tbody tr").filter(function() { console.log("%s, %d", $(this).find('td.Filename').text(), $(this).find('a').data('file-id'))})
test9.bin, 10
test8.bin, 8
test7.bin, 8
test6.bin, 7
test5.bin, 4
test4.bin, 5
test3.bin, 3
test2.bin, 3
test1.bin, 2
test0.bin, 1
```

This may be fixed with patch (which unblocks updating all file ids on upload not just for one file)...

```
--- a/var/httpd/htdocs/js/Core.UI.js   2021-11-08 18:53:34.088284466 +0100
+++ b/var/httpd/htdocs/js/Core.UI.js 2021-12-16 18:41:14.090199778 +0100
@@ -697,10 +697,6 @@

                                 $TargetObj = $ExistingItemObj.closest('tr');

-                                if ($TargetObj.find('a').data('file-id')) {
-                                    return;
-                                }
-
                                 $TargetObj
                                     .find('.Filetype')
                                     .text(Attachment.ContentType)
```

...but using file ids for identification of uploaded files is not race condition resistant; to provoke race condition apply patch

```
--- a/Kernel/Modules/AjaxAttachment.pm 2021-06-19 21:59:34.379471088 +0200
+++ b/Kernel/Modules/AjaxAttachment.pm       2021-12-16 18:38:28.943925139 +0100
@@ -56,6 +56,10 @@
             Param => 'Files',
         );

+        if ($UploadStuff{Filename} eq 'test01.bin') {
+            sleep(1);
+        }
+
         $UploadCacheObject->FormIDAddFile(
             FormID      => $Self->{FormID},
             Disposition => 'attachment',
@@ -67,6 +71,10 @@
             FormID => $Self->{FormID},
         );

+        if ($UploadStuff{Filename} eq 'test02.bin') {
+            sleep(2);
+        }
+
         my @AttachmentData;
```

and drop files test01.bin and test02.bin (same size) in one move; after upload there will be:

```
$(".AttachmentListContainer tbody tr").filter(function() { console.log("%s, %d", $(this).find('td.Filename').text(), $(this).find('a').data('file-id'))})
test02.bin, 1
test01.bin, 1
```

not

```
test02.bin, 2
test01.bin, 1
```

This problem was solved by removing unnecessarry FileID attribute from upload data struct to avoid any races that may occurr between filename list and fileid list. Only filename will be used to identify items added to upload cache.

(2)
Double clicking trash icon to remove attachment generates two requests to server and one will cause an error.

(3)
Web form is not removed from upload cache dir after saving or updating template.

(4)
Unused upload handling code removed.

Author-Change-Id: IB#1113065

+++ Kernel/System/Web/UploadCache/DB.pm
@@ -1,6 +1,7 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
 # Copyright (C) 2021 Znuny GmbH, https://znuny.org/
+# Copyright (C) 2021 Informatyka Boguslawski sp. z o.o. sp.k., http://www.ib.pl/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -122,7 +123,7 @@ sub FormIDAddFile {
 sub FormIDRemoveFile {
     my ( $Self, %Param ) = @_;
 
-    for my $Needed (qw(FormID FileID)) {
+    for my $Needed (qw(FormID Filename)) {
         if ( !$Param{$Needed} ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
                 Priority => 'error',
@@ -137,16 +138,22 @@ sub FormIDRemoveFile {
     # finish if files have been already removed by other process
     return if !@Index;
 
-    my $ID = $Param{FileID} - 1;
-    $Param{Filename} = $Index[$ID]->{Filename};
+    # Find and remove file with given filename; return success if
+    # file not found (to avoid error if user double clicks delete icon).
+    for my $File (@Index) {
+        if ($File->{Filename} eq $Param{Filename}) {
 
-    return if !$Kernel::OM->Get('Kernel::System::DB')->Do(
-        SQL => '
-            DELETE FROM web_upload_cache
-            WHERE form_id = ?
-                AND filename = ?',
-        Bind => [ \$Param{FormID}, \$Param{Filename} ],
-    );
+            return if !$Kernel::OM->Get('Kernel::System::DB')->Do(
+                SQL => '
+                    DELETE FROM web_upload_cache
+                    WHERE form_id = ?
+                        AND filename = ?',
+                Bind => [ \$Param{FormID}, \$Param{Filename} ],
+            );
+
+            last;
+        }
+    }
 
     return 1;
 }
@@ -154,7 +161,6 @@ sub FormIDRemoveFile {
 sub FormIDGetAllFilesData {
     my ( $Self, %Param ) = @_;
 
-    my $Counter = 0;
     my @Data;
     for my $Needed (qw(FormID)) {
         if ( !$Param{$Needed} ) {
@@ -180,7 +186,6 @@ sub FormIDGetAllFilesData {
     );
 
     while ( my @Row = $DBObject->FetchrowArray() ) {
-        $Counter++;
 
         # encode attachment if it's a postgresql backend!!!
         if ( !$DBObject->GetDatabaseFunction('DirectBlob') ) {
@@ -197,7 +202,6 @@ sub FormIDGetAllFilesData {
                 Filename    => $Row[0],
                 Filesize    => $Row[2],
                 Disposition => $Row[5],
-                FileID      => $Counter,
             }
         );
     }
@@ -208,7 +212,6 @@ sub FormIDGetAllFilesData {
 sub FormIDGetAllFilesMeta {
     my ( $Self, %Param ) = @_;
 
-    my $Counter = 0;
     my @Data;
     for my $Needed (qw(FormID)) {
         if ( !$Param{$Needed} ) {
@@ -233,7 +236,6 @@ sub FormIDGetAllFilesMeta {
     );
 
     while ( my @Row = $DBObject->FetchrowArray() ) {
-        $Counter++;
 
         # add the info
         push(
@@ -244,7 +246,6 @@ sub FormIDGetAllFilesMeta {
                 Filename    => $Row[0],
                 Filesize    => $Row[2],
                 Disposition => $Row[4],
-                FileID      => $Counter,
             }
         );
     }
